Update get all listings
