+++ .eslintrc
@@ -17,7 +17,7 @@
   "ignorePatterns": ["src/seeders/*.js", "src/migrations/*.js"],
   "rules": {
     "prettier/prettier": "error",
-    "max-len": ["error", { "code": 100 }],
+    "max-len": ["error", { "code": 120 }],
     "class-methods-use-this": 0,
     "import/no-absolute-path": 2,
     "import/no-extraneous-dependencies": 1,
